+++ akshare/__init__.py
@@ -2308,9 +2308,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.8.89 fix: fix sunrise_city_list interface
 1.8.90 fix: fix bond_info_detail_cm interface
 1.8.91 fix: fix sunrise_monthly interface
+1.8.92 fix: fix stock_institute_hold interface
 """
 
-__version__ = "1.8.91"
+__version__ = "1.8.92"
 __author__ = "AKFamily"
 
 import sys