+++ docs/changelog.md
@@ -49,6 +49,10 @@
 
 ## 更新说明
 
+1.8.92 fix: fix stock_institute_hold interface
+
+    1. 修复 stock_institute_hold 接口
+
 1.8.91 fix: fix sunrise_monthly interface
 
     1. 修复 sunrise_monthly 接口
@@ -1913,6 +1917,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.8.92 fix: fix stock_institute_hold interface
 
 1.8.91 fix: fix sunrise_monthly interface